refactor(api): syndicated article to new pattern
